Resolve many warnings, and coding guideline violations.
There seems to be some issue with editor formatting with
one of the authors of the code, which has made the code somewhat
difficult to read.

The first set of goals should be to remove all warnings, and while doing
that try to clean up some of the code formatting.