SAUTO requires a careful inspection and rework.
For the time being, USE_SAUTO is not defined by default. This means that
we do not need to compile the respective files into the libsystemc-clang
library.  The CMakeLists adjust that.

The source code must also use #ifdef and #endif pairs to isolate the
code that generates and dumps the SAUTO.  This has also been added.

The focus as of now is going to be primarily on sc_method processes.