A module can have sub-modules that are instances.
This resulted in quite a bit of re-work on the interfacing since I had
to store the list of instances.

There was a double free memory issue as well.  The key thing to remember
is that there is an incomplete ModuleDecl*, which propagates the
information common across different instances. Example includes ports.
Other instance-specific information is then later added.

The destructor for ModuleDecl should only delete the information that is
instance-specific.  The deletion of the incomplete ModuleDecl frees up
the common information.