Fix: Correct the index for the false path for if blocks.
The issue was somewhat specific as follows.

When there is an IF block, and the successor for the false is a block
that has been visited, the traversal reaches the IF block at the end,
and notices that there is no unvisited successor.  Consequently,
the false path index saved is the size of the currently visited
blocks, which is incorrect.  Instead, we need to find the false path
block in the currently visited blocks.

This should fix #368  and #373