False path information was being overwritten for each path traversal.
This is a partial fix. For now, for each path traversal, there is a
separate vector of true and false paths. This resulted in the
introduction of the all_paths_info_, which will then deprecate the
paths_info_.

TODO: The recursive merging of the path information is still needs to be
done.