Scratchllnl (#398)
* False path information was being overwritten for each path traversal.

This is a partial fix. For now, for each path traversal, there is a
separate vector of true and false paths. This resulted in the
introduction of the all_paths_info_, which will then deprecate the
paths_info_.

TODO: The recursive merging of the path information is still needs to be
done.

* Update the true/false path in the checks.

* Bug fix: wait SplitCFG block was identified as conditional.

This occured when the parent CFGBlock was a conditional. A wait SplitCFG
block is not a conditional block.

Tests are being updated too.

* Update test to reflect path info

---------

Co-authored-by: rseac <nospam@uwaterloo.ca>